+++ sql/updates/latest-dev.sql
@@ -297,3 +297,12 @@ CREATE FUNCTION @extschema@.add_reorder_policy(
 ) RETURNS INTEGER
 AS '@MODULE_PATHNAME@', 'ts_policy_reorder_add'
 LANGUAGE C VOLATILE;
+
+-- Trigger that blocks modifications on frozen chunks
+CREATE OR REPLACE FUNCTION _timescaledb_internal.frozen_chunk_modify_blocker() RETURNS trigger
+   LANGUAGE plpgsql STRICT AS
+$BODY$
+BEGIN
+     RAISE EXCEPTION 'Unable to modify frozen chunk %s', TG_TABLE_NAME;
+END;
+$BODY$ SET search_path TO pg_catalog, pg_temp;