+++ tsl/test/sql/chunk_utils_internal.sql
@@ -61,7 +61,19 @@ WHERE hypertable_name = 'hyper1' and hypertable_schema = 'test1'
 ORDER BY chunk_name LIMIT 1
 \gset
 
+-- Freeze and check trigger
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
 SELECT  _timescaledb_internal.freeze_chunk( :'CHNAME');
+
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
 
 SELECT * from test1.hyper1 ORDER BY 1;
 
@@ -75,6 +87,18 @@ DELETE FROM test1.hyper1 WHERE time = 20;
 DELETE FROM test1.hyper1 WHERE temp = 0.5;
 SELECT * from test1.hyper1 ORDER BY 1;
 
+-- TEST delete on unfrozen chunk should be possible
+UPDATE test1.hyper1 SET temp = 40 WHERE time = 20;
+
+-- Test with cast (chunk path pruning can not be done during query planning)
+UPDATE test1.hyper1 SET temp = 40 WHERE time = 20::text::float;
+
+-- TEST delete on unfrozen chunk should be possible
+DELETE FROM test1.hyper1 WHERE time = 20;
+
+-- Test with cast
+DELETE FROM test1.hyper1 WHERE time = 20::text::float;
+
 -- TEST inserts into a frozen chunk fails
 INSERT INTO test1.hyper1 VALUES ( 11, 11);
 \set ON_ERROR_STOP 1
@@ -87,11 +111,29 @@ SELECT * from test1.hyper1 ORDER BY 1;
 SELECT table_name, status
 FROM _timescaledb_catalog.chunk WHERE table_name = :'CHUNK_NAME';
 
+-- Check trigger
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
 SELECT  _timescaledb_internal.unfreeze_chunk( :'CHNAME');
+
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
 
 --verify status in catalog
 SELECT table_name, status
 FROM _timescaledb_catalog.chunk WHERE table_name = :'CHUNK_NAME';
+
+-- Test update works after unfreeze
+UPDATE test1.hyper1 SET temp = 40;
+
+-- Test delete works after unfreeze
+DELETE FROM test1.hyper1;
+
 --unfreezing again works
 SELECT  _timescaledb_internal.unfreeze_chunk( :'CHNAME');
 SELECT  _timescaledb_internal.drop_chunk( :'CHNAME');
@@ -169,6 +211,10 @@ SELECT  compress_chunk( :'CHNAME');
 \set ON_ERROR_STOP 0
 SELECT _timescaledb_internal.drop_chunk(:'CHNAME');
 \set ON_ERROR_STOP 1
+
+-- Prepare table for CAGG tests
+TRUNCATE test1.hyper1;
+INSERT INTO test1.hyper1(time, temp) values(30, 0.5), (31, 31);
 
 --TEST drop_chunk in the presence of caggs. Does not affect cagg data
 CREATE OR REPLACE FUNCTION hyper_dummy_now() RETURNS BIGINT
@@ -268,7 +314,8 @@ SELECT * FROM ht_try ORDER BY 1;
 
 SELECT relname, relowner::regrole FROM pg_class
 WHERE relname in ( select chunk_name FROM chunk_view 
-                   WHERE hypertable_name = 'ht_try' );
+                   WHERE hypertable_name = 'ht_try' )
+ORDER BY relname;
 
 SELECT inhrelid::regclass
 FROM pg_inherits WHERE inhparent = 'ht_try'::regclass ORDER BY 1;
@@ -394,6 +441,125 @@ FROM chunk_view
 WHERE hypertable_name = 'hyper_constr'
 ORDER BY chunk_name;
 
+----- TESTS for copy into frozen chunk ------------
+\c :TEST_DBNAME :ROLE_DEFAULT_PERM_USER
+CREATE TABLE test1.copy_test (
+    "time" timestamptz NOT NULL,
+    "value" double precision NOT NULL
+);
+
+SELECT create_hypertable('test1.copy_test', 'time', chunk_time_interval => interval '1 day');
+
+COPY test1.copy_test FROM STDIN DELIMITER ',';
+2020-01-01 01:10:00+01,1
+2021-01-01 01:10:00+01,1
+\.
+
+-- Freeze one of the chunks
+SELECT chunk_schema || '.' ||  chunk_name as "COPY_CHNAME", chunk_name as "COPY_CHUNK_NAME"
+FROM timescaledb_information.chunks
+WHERE hypertable_name = 'copy_test' and hypertable_schema = 'test1'
+ORDER BY chunk_name LIMIT 1
+\gset
+
+SELECT _timescaledb_internal.freeze_chunk( :'COPY_CHNAME');
+
+-- Check state
+SELECT table_name, status
+FROM _timescaledb_catalog.chunk WHERE table_name = :'COPY_CHUNK_NAME';
+
+\set ON_ERROR_STOP 0
+-- Copy should fail because one of che chunks is frozen
+COPY test1.copy_test FROM STDIN DELIMITER ',';
+2020-01-01 01:10:00+01,1
+2021-01-01 01:10:00+01,1
+\.
+\set ON_ERROR_STOP 1
+
+-- Count existing rows
+SELECT COUNT(*) FROM test1.copy_test;
+
+-- Test dump & restore
+\c postgres :ROLE_SUPERUSER
+\! utils/pg_dump_aux_dump.sh dump/pg_dump.sql
+
+\c :TEST_DBNAME
+
+-- Make sure tables was droped by pg_dump_aux_dump.sh
+\set ON_ERROR_STOP 0
+SELECT * FROM test1.copy_test;
+\set ON_ERROR_STOP 1
+
+SET client_min_messages = ERROR;
+CREATE EXTENSION timescaledb CASCADE;
+RESET client_min_messages;
+
+--\! cp dump/pg_dump.sql /tmp/dump.sql
+SELECT timescaledb_pre_restore();
+\! utils/pg_dump_aux_restore.sh dump/pg_dump.sql
+SELECT timescaledb_post_restore();
+SELECT _timescaledb_internal.stop_background_workers();
+
+\c :TEST_DBNAME :ROLE_DEFAULT_PERM_USER
+
+-- Make sure the chunk is still frozen
+-- Check trigger
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'COPY_CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
+-- Check state
+SELECT table_name, status
+FROM _timescaledb_catalog.chunk WHERE table_name = :'COPY_CHUNK_NAME';
+
+\set ON_ERROR_STOP 0
+-- Copy should fail because one of che chunks is frozen
+COPY test1.copy_test FROM STDIN DELIMITER ',';
+2020-01-01 01:10:00+01,1
+2021-01-01 01:10:00+01,1
+\.
+\set ON_ERROR_STOP 1
+
+-- Count existing rows
+SELECT COUNT(*) FROM test1.copy_test;
+
+-- Check unfreeze restored chunk
+SELECT _timescaledb_internal.unfreeze_chunk( :'COPY_CHNAME');
+
+-- Check trigger
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'COPY_CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
+-- Check state
+SELECT table_name, status
+FROM _timescaledb_catalog.chunk WHERE table_name = :'COPY_CHUNK_NAME';
+
+-- Copy should work now 
+COPY test1.copy_test FROM STDIN DELIMITER ',';
+2020-01-01 01:10:00+01,1
+2021-01-01 01:10:00+01,1
+\.
+
+-- Test that unfreeze works even if somebody has dropped the block trigger
+SELECT _timescaledb_internal.freeze_chunk( :'COPY_CHNAME');
+
+SELECT table_name, status
+FROM _timescaledb_catalog.chunk WHERE table_name = :'COPY_CHUNK_NAME';
+
+DROP TRIGGER frozen_chunk_modify_blocker ON :COPY_CHNAME;
+
+SELECT trigger_name, event_manipulation, action_statement, action_orientation, action_timing
+FROM information_schema.triggers WHERE trigger_catalog = 'db_chunk_utils_internal' AND
+event_object_table = :'COPY_CHUNK_NAME'
+ORDER BY event_object_table, trigger_name, event_manipulation;
+
+SELECT _timescaledb_internal.unfreeze_chunk( :'COPY_CHNAME');
+
+SELECT table_name, status
+FROM _timescaledb_catalog.chunk WHERE table_name = :'COPY_CHUNK_NAME';
 
 -- clean up databases created
 \c :TEST_DBNAME :ROLE_SUPERUSER